Add a git-ctx command
This relies on the same configuration format than git-tui, and allows picking
repositories and branches easily, à la kubectx. git-ctx is meant to be exec'ed
so as to not open a new shell but replace the existing one. This can easily
achieved by aliasing `gctx` to `exec git-ctx`. Note that this is only intended
to and tested on the Fish shell.